backend/requirements/base.txt
Django>=5.0,<6.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.1
django-cors-headers>=4.3.1
django-filter>=23.5
django-ratelimit>=4.1.0
drf-yasg>=1.21.7
psycopg2-binary>=2.9.9
redis>=5.0.1
celery>=5.3.6
gunicorn>=21.2.0
Pillow>=10.2.0
python-dotenv>=1.0.1
cryptography>=41.0.7
django-extensions>=3.2.3
pytest-django>=4.7.0
twilio
djangorestframework-api-key
